Refactor all Mastermind messages into MastermindDialog class
